Add test for azure secrets
